+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - **Custom Color Gradient** custom gradient defined by a start and end color
+- **SuggestionSupplier** for getting heatmap panel recommended as suggested visualization
 
 ### Changed
 
@@ -18,7 +19,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 TODO:
 - data links option + menu
 - option to use data timeframe
-- add suggestions supplier
 
 ## [1.1.1] - 2026-05-03